add validate file
